Finally nailed MessageCell.
The key was to have two views, one for each of the balloons. Both are loaded when a cell is initialized, Then I just turn on the one I need for the display of the cell. This is because cells are cached, so while I might need my balloon now, the next time the cell is fetched from the cache it might need to display your balloon. Phew! This also allows me to cache both images up-front, rather than reloading the UIImage and setting it to be resizeable each time a cell is displayed.