Rewrote from scratch.
Following Loren Brichter's [recommendations](http://blog.atebits.com/2008/12/fast-scrolling-in-tweetie-with-uitableview/) for speed based on his work with Tweetie, I threw everything out and started over. Took a while, but the last part, the BuzzerViewController, went a lot faster than before. The layout is coming right along. Most of the original stuff is still in place visually, but the code is much simpler, using Core Graphics to do all the drawing, and has no nibs at all. I might end up adding some to handle the dialogs, but maybe not. We'll see. Anyway, very happy with the progress on the UI now.