Fix crash.
I'm still not clear on when I should release objects and when I should not. Need to read up on that. Anyway, the main screen no longer crashes when I scroll a few times, simply by removing `[buzzer release]` from `setBuzzer:` in `BuzzerCell`.

The rest of the commit is just code reformatting in `RootViewController`.