Added `SQLMigrator`.
It runs every time the app starts, checking the current schema version, and runs migration scripts as appropriate. It uses the `schema_version` pragma to track what migration has been run, which is nice and clean. Seems to be mostly working, although the home screen crashes as soon as I try to do anything to it. I'll fix that tomorrow.