Eliminate memory leaks.
Used the Instrumentation app. Works great. Main leak, after my own fiddling, was the `icons` variable in `MessageModel`. I changed it to a static variable, since it's just a lookup table, and the leaks disappeared. W00t!'

My own fiddling was more careful handling of the accessors in the cell classes and auto-releasing the array returned by the `select` methods in the models, as well as a missing `release` in the `RootViewController`.