Factor all icon-handling code into `IconFinder`.
It now has just two methods: `loadForEmail` and `updateThumbsForEmails`. The former is now used by `BuzzerModel` to load an icon for an email address, and will load the file from the documents directory if it exists, and look for it in the address book if it doesn't. The latter isn't used yet, but will be set on a timer to periodically reload any icons from the address book that might have changed.

These two methods rely on the private `findForEmails` method, which does the actual address book searching. It could be extended in the future to search elsewhere, such as Gravatar, Facebook, etc.

In the process, I removed `UIImage+SaveTo`, because it was much too specific to go into the `UIImage`class.

Still to do: Allow `loadForEmail` to run asynchronously if it needs to search the address book for an image. I might have to do that elsewhere, I'm not sure.