Made image loading async.
It's handled by the view. Currently icons are disabled in the `MessageCell`. Issues still to be addressed:

* `IconFinder` uses `imageWithContentsOfFile:` to load icons from the Documents folder, but that method doesn't cache the image like `imageNamed:`, and `imageNamed:` only works for the application bundle, not the standbox. So I need to have a better way to cache images. It may not be necessary in the root view, as each correspondent is displayed only once. But I'm definitely going to want it in the buzzer view.

* The cell view may not be the right place for this stuff. Perhaps better would be to do it in the controller and store the fetched icon in the model? That would allow the icons to be cached with the model, which would then extend to the buzzer view, which would be nice.

* Otherwise, I'll probably have to add some sort of independent cache to `IconFinder` using an `NSMutableDictionary` object and add a method to detect memory warnings and clear it of all objects. Will have to look at what that would take.

Once that's working, I'll add the code to periodically update the thumbnail images from the address book, to take advantage of changes in the address book. I suspect that'll be put on a timer, though maybe I can put stuff in the `tmp` directory in the sandbox? Might that directory be periodically cleared by the system? Will need to look.