Switch to relative dates.
This is done in the `BuzzerCell` class. The `most_recent` view now returns the `sent_at` time in epoch seconds, and the new method, `relativeTimeSince:`, converts it to a nicely formatted relative time.

I'm suspicious of how well it works, though, so I've also started adding support for unit tests. More on that next.