Add `NSDate+HumanInterval` and convert to it.
This category adds the method `humanIntervalSinceNow`, which has a shortened relative date for humans to read. I've also added tests for it in `HumanIntervalTest`. Good thing, too, because there were a bunch of bugs.

I've also reverted the `most_recent` view to displaying a regular date, and chagned the query against it to turn it into an epoch date. I'm now turning that epoch date into an `NSDate` object so that I can then just take advantage of `humanIntervalSinceNow`. It also means that the `when` object of `BuzzerModel` is an `NSDate` object, so it's more correct. (Whether the time zone is right remains to be seen).