Format timestamp in mesage bubbles.
Figured out how to assume the date stored in SQLite is UTC and convert to the local time for display. Then changed the `sent` attribute of `MessageModel` to an `NSDate` objet and added formatting of the timestamp of individual messages in bubbles using `NSDateFormatter` and centered it in the cell.