Some IconFinder fixes.
I reset things and discovered it wasn't finding images anymore. This is because I had a `break` instead of a `continue`. D'oh!

When I recreated my address book entries, I added a lowercased address when the address in the database is mixed case. The icon wasn't found. So I tweaked things again so that the emails passed to `-findForEmails:` are now suckec into another array in lowercase, and then the address book entries are lowercased. So now the images can be found in the address book without regard to case. Yay!