Add email configuration.
This is just a start, but it's storing the email configuration in the user defaults and setting up a background process to find the icon when the email address is changed. I've also updated `BuzzerViewController` so that it stores my icon and your icon, and assigns it as appropriate to each message cell. Might end up moving that into the `MessageCell` class; I'm not sure.

The implementation of the configuration stuff is copied from my original experimental implementation, and is now a table; I might change it later.

Oh, and I found a typo in the test data where a message that was supposed to be from me was showing as from Rick LePage.