Start adding address detail management.
The new class, `AddressViewController`, does the work. In the process, I discovered a better way to make glass button backgrounds (using a private API from Apple, but just to generate the graphics), so use those graphics. I also found a nice character to use for the reordering button (three lines, like the reordering icons themselves), so I'm using that. It's just a Unicode character, which is nice. Gotta keep remembering that. And finally, I tweaked the color of the settings fields text.