Add some actions on confirmation.
Basically we'll send a request or a confirmation. So handling the dynamics of the UI for that. Got it all working except that the config table view needs to be reloaded before popping off the confirming view. But the confirm code field and button are dynamically added after a successful request. Also, making use of MBProgressHUD to display the status messages during and after the requests. It looks good.