Update config view when address view popped off.
Any time an address view is dismissed, we re-read the data for
the config view. This is so that any changes in the address view
are reflected in the config view.

Doing this is simple, but took forever to figure out:

* Move the population of the config view data from `viewDidLoad`
  to `ViewWillAppear:`, as the latter is called every time the
  view displays, while the former only the first time it
  displays.
* Grab the config view in the addres view's `ViewWillDisappear:`
  method and tell it to reload its data.

So yeah, there could be times when no real changes are made, but
that should be fairly rare, and this is a pretty cheap series of
operations anyway.