Fix crash when deleting just-added address.
The table was reloading before the addresses were reloaded. So
I've changed it to reload the addresses from the address
controller instead of the config controller.

While at it, I moved creation of a new address to when the
request is submitted, not before. I've also eliminated a
superfluous setting of the accessory type in the config
controller.