Restore primary email address setting.
I've updated `RootViewController` so that, when the config
controller is dismissed, a new background task is started where
all the confirmed email addresses are found, the first one saved
as the primary, and then they're all passed off to the
`updateThumbsForEmails:` method of `IconFinder`. This refreshes
all the icons for all email addresses. It has been there a while,
but was not yet used. It had a simple bug, easily fixed.

While at it, I changed the setup of the address view so that the
email field becomes the first responder when it loads if there is
not already an address.