Try harder to find my icon.
Add `findAmongEmails:cacheFor:` to `IconFinder`. It tries to find
an icon in a list of icons and returns the first it finds,
caching it for `cacheFor:`. If it's not found, and there isn't some
image present already, it caches the default icon.

Then switched `updateMyIcons:` in `RootViewController` to create
an array of all email addresess, figure out which is the primary,
and then pass them off to `findAmongEmails:cacheFor:`.