Add address picker when composing.
This is only when clicking the "new message" icon from the root view. It pops
up a `ABPeoplePickerNavigationController`, the user can select a recipient
then select an email address, and then it fades to the
`ComposeViewController`. The latter appears as a modal view of the picker,
thus appearing to replace the picker. I also created
`ComposeViewControllerDelegate` so that the root view controller can have the
control to dismiss its modal dialog, which is the picker, and that dismisses
both the picker and the composer at the same time, making it very clean.

Next up, I need to make it so that the keyboard just appears in the compose
window, without animation, and then to add the compose window to the buzzer
controller. Both will need to take the correspondent information and do the
right thing with it (using a typedef for that for now).