Add Three20-based composer.
Seems to work well with the proxy, except that I can't figure out how to free
the proxy from memory when I'm done. So I have a memory leak at the moment,
but it looks good!

Once I figure out how to deal with the leak, I'll need to get the addressing
stuff working well and the sending progress.